fix: Fix Footer Navigation and add MainScreen. Remove bottom_nav_bar file. detail: Now the MainScreen contatins: Home, Chatbot, Camera, Profile and don't need to navigate by router. There was a bit change in router file.
